services/llm-multimodal/app.py
"""
LLM Multimodal service for Smartrobe.

Extracts contextual attributes using multimodal large language models.
Combines vision and text understanding for sophisticated attribute reasoning.
Currently implements season appropriateness detection.
"""

import asyncio
import random
import time
from typing import Any

from PIL import Image

from shared.base_service import CapabilityService
from shared.schemas import Season


class LLMMultimodalService(CapabilityService):
    """LLM Multimodal service for contextual attribute extraction."""

    SUPPORTED_ATTRIBUTES = {
        "season": Season,
    }
    SERVICE_TYPE = "llm"

    def __init__(self):
        super().__init__("llm_multimodal", "llm", "1.0.0")

    async def _initialize_service(self) -> None:
        """Initialize LLM multimodal service."""
        # In a real implementation, this would:
        # - Initialize connection to multimodal LLM API (GPT-4V, Claude 3, etc.)
        # - Load prompt templates for clothing analysis
        # - Set up image encoding/processing pipeline
        # - Configure retry and rate limiting
        # - Initialize vision-language understanding components
        await asyncio.sleep(0.6)
        
        self.logger.info("LLM multimodal service initialized successfully")
        self.set_health_detail("llm_api_connected", True)
        self.set_health_detail("vision_encoder_loaded", True)
        self.set_health_detail("prompt_templates_loaded", True)
        self.set_health_detail("multimodal_pipeline_ready", True)
        self.set_health_detail("model_type", "mock_gpt4_vision")

    async def _cleanup_service(self) -> None:
        """Cleanup LLM multimodal service."""
        self.logger.info("LLM multimodal service cleaned up")

    async def extract_single_attribute(
        self, request_id: str, attribute_name: str, image_paths: list[str]
    ) -> tuple[Any, float]:
        """
        Extract a single attribute using multimodal LLM reasoning.

        Args:
            request_id: Unique request identifier
            attribute_name: Name of the attribute to extract
            image_paths: List of paths to images

        Returns:
            Tuple of (attribute_value, confidence_score)
        """
        self.logger.info(
            f"Extracting {attribute_name} using multimodal LLM",
            request_id=request_id,
            attribute=attribute_name,
            image_count=len(image_paths),
        )

        if attribute_name == "season":
            return await self._extract_season(request_id, image_paths)
        else:
            raise ValueError(f"Unsupported attribute: {attribute_name}")

    async def _extract_season(self, request_id: str, image_paths: list[str]) -> tuple[Season, float]:
        """
        Extract seasonal appropriateness using multimodal LLM reasoning.

        In a real implementation, this would:
        1. Encode images using vision transformer
        2. Generate detailed image descriptions
        3. Analyze fabric weight, colors, and styling
        4. Consider layering possibilities and coverage
        5. Apply seasonal fashion knowledge
        6. Reason about weather appropriateness
        7. Generate confidence based on reasoning consistency

        Args:
            request_id: Unique request identifier
            image_paths: List of paths to images

        Returns:
            Tuple of (Season, confidence_score)
        """
        # Simulate multimodal LLM processing time (slower than heuristics)
        processing_time = random.uniform(0.8, 1.5)
        await asyncio.sleep(processing_time)

        # Mock multimodal reasoning with seasonal logic
        # In reality, this would analyze fabric, colors, styling, coverage
        season_reasoning = {
            Season.SUMMER: {
                "probability": 0.25,
                "indicators": ["lightweight fabrics", "bright colors", "short sleeves", "minimal coverage"],
                "confidence_range": (0.85, 0.95)
            },
            Season.WINTER: {
                "probability": 0.20,
                "indicators": ["heavy fabrics", "dark colors", "long sleeves", "layering potential"],
                "confidence_range": (0.88, 0.96)
            },
            Season.SPRING: {
                "probability": 0.22,
                "indicators": ["medium weight", "pastel colors", "transitional pieces"],
                "confidence_range": (0.75, 0.90)
            },
            Season.FALL: {
                "probability": 0.23,
                "indicators": ["medium-heavy fabrics", "earth tones", "layering friendly"],
                "confidence_range": (0.78, 0.92)
            },
            Season.ALL_SEASON: {
                "probability": 0.10,
                "indicators": ["versatile design", "neutral colors", "adaptable styling"],
                "confidence_range": (0.70, 0.85)
            }
        }

        # Select season based on probabilities
        seasons = list(season_reasoning.keys())
        weights = [data["probability"] for data in season_reasoning.values()]
        selected_season = random.choices(seasons, weights=weights)[0]

        # Generate confidence based on reasoning strength
        confidence_range = season_reasoning[selected_season]["confidence_range"]
        confidence = random.uniform(*confidence_range)

        # Simulate reasoning process (would be actual LLM reasoning)
        reasoning_indicators = season_reasoning[selected_season]["indicators"]
        selected_indicators = random.sample(reasoning_indicators, min(2, len(reasoning_indicators)))

        self.logger.info(
            "Season extraction completed",
            request_id=request_id,
            detected_season=selected_season.value,
            confidence=confidence,
            reasoning_indicators=selected_indicators,
            processing_time_ms=int(processing_time * 1000),
        )

        return selected_season, confidence

    async def _check_service_health(self) -> bool:
        """Check LLM multimodal service health."""
        try:
            # Test multimodal LLM API connection and basic reasoning
            test_start = time.time()
            await asyncio.sleep(0.05)  # Simulate API health check
            
            health_check_time = time.time() - test_start
            self.set_health_detail("last_health_check_ms", int(health_check_time * 1000))
            self.set_health_detail("api_response_time_ms", int(health_check_time * 1000))
            self.set_health_detail("multimodal_reasoning_ready", True)
            self.set_health_detail("vision_understanding_ready", True)
            
            return True
        except Exception as e:
            self.logger.error("LLM multimodal service health check failed", error=str(e))
            return False


def create_app():
    """Create the LLM multimodal FastAPI application."""
    service = LLMMultimodalService()
    return service.app


# Create app instance for uvicorn
app = create_app()


# For development/testing
if __name__ == "__main__":
    service = LLMMultimodalService()
    service.run()
